+++ scripts/duc_noip_install.sh
@@ -4,6 +4,13 @@
 # THE SOFTWARE IS PROVIDED "AS IS"
 # Read the included LICENSE file for more information
 
+# Exit codes:
+# 1  - The Dynamic Update Client is already installed.
+# 2  - No internet
+# 3  - The standard input was redirected
+# 4  - Failed to get root privileges
+# 10 - Something went wrong during the installation
+
 # Check if the DUC is already installed
 if which noip2 &>/dev/null; then
 	printf "No-IP's DUC is already installed\n" >&2
@@ -34,9 +41,12 @@ Examples:
 	./%s -p /usr -n\n" "$(basename "$0")" "$(basename "$0")"
 }
 
-# Process options
+# Prepare variables
 integrate_desktop=yes
 prefix="$HOME/.local"
+[ $(id -u) == 0 ] && prefix="/usr/local"
+
+# Process options
 while [ -n "$1" ]; do
 	case "$1" in
 		-n) integrate_desktop=no ;;
@@ -53,24 +63,23 @@ printf "Installing \e[01;36mNo-IP's Dynamic Update Client\e[00m (DUC):
 This script may open a text editor to work around an issue when typing your
 password, see the README file for more information.\n"
 
-# Download the program
+# Download and install the program
+status=ok
 cd /usr/local/src/
-sudo wget -q http://www.noip.com/client/linux/noip-duc-linux.tar.gz
+sudo wget -q http://www.noip.com/client/linux/noip-duc-linux.tar.gz || status=bad
 
 # Unpack, install and configure
-installed=no
-if sudo tar -zxf noip*.tar.gz ; then
-	sudo rm noip.tar.gz
+if [ "$status" = "ok" ] && sudo tar -zxf noip*.tar.gz ; then
+	sudo rm noip*.tar.gz
 	cd noip*/
 	gedit </dev/null &>/dev/null &
-	sudo make install
-	installed=yes
+	sudo make install || status=bad
+else
+	status=bad
 fi
 
 # Integrate with the user's desktop
-if [ "$integrate_desktop" = "yes" -a "$installed" = "yes" ]; then
-	cd "$HOME"
-
+if [ "$integrate_desktop" = "yes" -a "$status" = "ok" ]; then
 	# Create relevant directories and files
 	target="$prefix/bin"
 	entry="$prefix/share/applications"
@@ -83,7 +92,8 @@ if [ "$integrate_desktop" = "yes" -a "$installed" = "yes" ]; then
 	icon="$icon/noip-duc.png"
 
 	# Make the "binary" (it's actually a bash script) file
-	printf "%s\n" "#!/bin/bash
+	cat <<EOF > "$target" || status=bad
+#!/bin/bash
 # Target file to help you manage No-IP\'s DUC
 
 # File created by duc_noip_install.sh script.
@@ -93,58 +103,61 @@ if [ "$integrate_desktop" = "yes" -a "$installed" = "yes" ]; then
 action=start
 
 USAGE_MSG () {
-	printf \"Usage:
-	./%s (-k)\\n\"
+	printf "Usage:
+	./%s (-k)\n"
 }
 
-while [ -n \"\$1\" ]; do
-	case \"\$1\" in
+# Acquire root privileges now
+sudo echo >/dev/null || exit 1
+
+while [ -n "\$1" ]; do
+	case "\$1" in
 		-k) action=kill                   ;;
 		-h|--help) USAGE_MSG >&2 ; exit 0 ;;
 		*)
-		printf \"option \\e[01m%s\\e[00m not recognized\\n\" \"\$1\" >&2
+		printf "option \e[01m%s\e[00m not recognized\n" "\$1" >&2
 		USAGE_MSG >&2
 		exit 1
 		;;
 esac; shift; done
 
-if [ \"\$action\" = \"start\" ]; then
-	printf \"root privileges are needed for this operation:\\n\"
-	if sudo /usr/local/bin/noip2 ; then
-		printf \"Process started\\n\"
-		sudo /usr/local/bin/noip2 -S
+if [ "\$action" = "start" ]; then
+	if sudo noip2 ; then
+		printf "Process started\n"
+		sudo noip2 -S
 		OUT=0
 	else
 		OUT=1
 	fi
-elif [ \"\$action\" = \"kill\" ]; then
+elif [ "\$action" = "kill" ]; then
 	PID=\$(ps ax | grep noip2 | grep -v grep | awk '{print \$1}' 2>/dev/null)
-	if [ -n \"\$PID\" ]; then
-		printf \"root privileges are needed for this operation:\\n\"
-		if sudo kill -9 \"\$PID\" ; then
-			printf \"Process successfully stopped\\n\"
+	if [ -n "\$PID" ]; then
+		if sudo kill -9 "\$PID" ; then
+			printf "Process successfully stopped\n"
 			OUT=0
 		fi
 	else
-		printf \"Process id could not be foud\\nMay not be running\\n\" >&2
+		printf "Process id could not be foud\nMay not be running\n" >&2
 		OUT=1
 	fi
 fi
 
 sleep 2
-exit \$OUT\\n" | tee "$target" >/dev/null
+exit \$OUT
+EOF
 	chmod 755 "$target"
 
 	# Download the icon
-	wget -q https://www.dropbox.com/s/g55zl9q9uc2a1pw/noiplogo.png?dl=1 -O "$icon"
+	wget -q https://www.dropbox.com/s/g55zl9q9uc2a1pw/noiplogo.png?dl=1 -O "$icon" || status=bad
 	chmod 644 "$icon"
 
 	# Make the desktop entry file
-	printf "%s\n" "[Desktop Entry]
+	cat <<EOF > "$entry" || status=bad
+[Desktop Entry]
 Name=No-Ip DUC
 Comment=Start the dynamic update client.
 GenericName=DUC;
-Exec=$target
+Exec=noip-assist
 Icon=noip-duc
 Type=Application
 StartupNotify=false
@@ -153,13 +166,18 @@ Categories=Network;Server;
 Actions=open-killing;
 Keywords=Network;Minecraft;Server;
 
-X-Desktop-File-Install-Version=0.24
-
 [Desktop Action open-killing]
 Name=Open to end process
-Exec=$target -k\\n" | tee "$entry" >/dev/null
+Exec=noip-assist -k
+EOF
 	chmod 644 "$entry"
 fi
 
-printf "\e[01;32mFinished!\e[00m The DUC installation script is done\n"
+if [ "$status" = "ok" ]; then
+	printf "\e[01;32mSuccess!\e[00m The script ran successfully, and the DUC is installed"
+	exit 0
+else
+	printf "\e[01;31mFailed!\e[00m The script finished, but something went wrong in the process"
+	exit 10
+fi
 # Thanks for downloading, and enjoy!